Images uplaod Issue - part 2 (User can edit or delte images uploaded by him and not others)

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -93,13 +93,28 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        // Logged in User can edit images only the ones which were uploaded by him
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedUserId = user.getId();
         Image image = imageService.getImage(imageId);
+        Integer imageUserId = image.getUser().getId();
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(loggedUserId == imageUserId ){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else{
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("image",image);
+            model.addAttribute("editError",error);
+            return "images/image";
+
+        }
+
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -141,9 +156,24 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        // Logged in User can delete images only the ones which were uploaded by him
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedUserId = user.getId();
+        Image image = imageService.getImage(imageId);
+        Integer imageUserId = image.getUser().getId();
+
+        if(loggedUserId == imageUserId ){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("image",image);
+            model.addAttribute("deleteError",error);
+            return "images/image";
+
+        }
     }
